Complete-changed the game type
I had tried doing what I originally said I was going to do but I had a lot of problems come up. I would fix one problem and get another.  I decided to scrap that idea and do something else. Red= -15 pts, Yellow= +1, and the flys are like the redblood cells. I used WBC as a platform and didn't really change their names because it made it easier and so less problems would pop up. I hope that this is acceptable. The goal of this is to get the highest score you can and avoid the red circles.